requirements.txt
# Core dependencies
openai>=1.0.0
python-dotenv>=0.19.0

# Testing dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-mock>=3.10.0

# Add these to requirements.txt
google-auth-oauthlib>=0.4.6
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0

# Calendar dependencies
icalendar>=5.0.0
pytz>=2023.3 